🚑 fix : 블랙리스트 추가, 멤버 삭제 로직 변경 , TTL 6시간으로 늘림

+++ src/main/java/org/team1/keyduck/member/service/MemberService.java
@@ -4,9 +4,13 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.team1.keyduck.auction.entity.AuctionStatus;
+import org.team1.keyduck.auction.repository.AuctionRepository;
+import org.team1.keyduck.auth.service.JwtBlacklistService;
 import org.team1.keyduck.common.exception.DataNotFoundException;
 import org.team1.keyduck.common.exception.DataNotMatchException;
 import org.team1.keyduck.common.exception.ErrorCode;
+import org.team1.keyduck.common.exception.OperationNotAllowedException;
 import org.team1.keyduck.member.dto.request.MemberUpdatePasswordRequestDto;
 import org.team1.keyduck.member.dto.request.MemberUpdateRequestDto;
 import org.team1.keyduck.member.dto.response.MemberReadResponseDto;
@@ -20,12 +24,14 @@ public class MemberService {
 
     private final MemberRepository memberRepository;
     private final PasswordEncoder passwordEncoder;
+    private final AuctionRepository auctionRepository;
+    private final JwtBlacklistService jwtBlacklistService;
 
     @Transactional
     public MemberUpdateResponseDto updateMember(MemberUpdateRequestDto requestDto, Long id) {
 
-        Member member = memberRepository.findById(id).orElseThrow(() -> new DataNotFoundException(
-                ErrorCode.NOT_FOUND_USER, "멤버"));
+        Member member = memberRepository.findById(id)
+                .orElseThrow(() -> new DataNotFoundException(ErrorCode.NOT_FOUND_USER, "멤버"));
 
         member.updateMember(requestDto);
 
@@ -35,8 +41,8 @@ public class MemberService {
     @Transactional
     public void updatePassword(MemberUpdatePasswordRequestDto requestDto, Long id) {
 
-        Member member = memberRepository.findById(id).orElseThrow(() -> new DataNotFoundException(
-                ErrorCode.NOT_FOUND_USER, "멤버"));
+        Member member = memberRepository.findById(id)
+                .orElseThrow(() -> new DataNotFoundException(ErrorCode.NOT_FOUND_USER, "멤버"));
 
         if (!passwordEncoder.matches(requestDto.getBeforePassword(), member.getPassword())) {
             throw new DataNotMatchException(ErrorCode.INVALID_DATA_VALUE, "비밀번호");
@@ -48,17 +54,28 @@ public class MemberService {
     }
 
     @Transactional
-    public void deleteMember(Long id) {
-        Member member = memberRepository.findById(id).orElseThrow(() -> new DataNotFoundException(
-                ErrorCode.NOT_FOUND_USER, "멤버"));
+    public void deleteMember(Long id, String token) {
+        Member member = memberRepository.findByIdAndIsDeleted(id, false);
+
+        if (member == null) {
+            throw new DataNotFoundException(ErrorCode.NOT_FOUND_USER, "멤버");
+        }
+
+        //현재 진행중인 경매가 있으면 탈퇴 불가능
+        if ((auctionRepository.existsByKeyboard_Member_IdAndAuctionStatus(id,
+                AuctionStatus.IN_PROGRESS))) {
+            throw new OperationNotAllowedException(ErrorCode.DELETE_FAIL_AUCTION_IN_PROGRESS, null);
+        }
+
+        jwtBlacklistService.addToBlacklist(token);
 
         member.deleteMember();
     }
 
     @Transactional(readOnly = true)
     public MemberReadResponseDto getMember(Long id) {
-        Member member = memberRepository.findById(id).orElseThrow(() -> new DataNotFoundException(
-                ErrorCode.NOT_FOUND_USER, "멤버"));
+        Member member = memberRepository.findById(id)
+                .orElseThrow(() -> new DataNotFoundException(ErrorCode.NOT_FOUND_USER, "멤버"));
 
         return MemberReadResponseDto.of(member);
     }
